Cleaned
Very nicely polished, saves header file even. Only self-exit would be on errors with user input that weren't caught, like the lack of a check if a proper header (01_01_BLOT_01) was entered, and if a user inputs a non-de blob 2 DS file. The thing doesn't even check if the file inputted is an str, but come on, it's a multi-use program, even if it is supposed to edit db2 dialogue, like, the thing sucks dude.